Refactor and add tests to Ast- and ElementPrinter
